chore: adopt first unary interceptors for stream calls
