+++ libs/common/grpc.go
@@ -1,36 +1,26 @@
 package common
 
 import (
-	"common/locale"
+	"common/hwgrpc"
 	"context"
 	"encoding/base64"
 	"errors"
 	"fmt"
 	"github.com/dapr/dapr/pkg/proto/runtime/v1"
 	daprd "github.com/dapr/go-sdk/service/grpc"
-	"github.com/go-playground/validator/v10"
-	"github.com/golang/protobuf/proto" //nolint:staticcheck
+	//nolint:staticcheck
 	"github.com/google/uuid"
 	prometheusGrpcProvider "github.com/grpc-ecosystem/go-grpc-middleware/providers/prometheus"
 	"github.com/grpc-ecosystem/go-grpc-middleware/v2/interceptors/auth"
 	"github.com/grpc-ecosystem/go-grpc-middleware/v2/metadata"
-	"github.com/rs/zerolog"
 	zlog "github.com/rs/zerolog/log"
 	"go.opentelemetry.io/contrib/instrumentation/google.golang.org/grpc/otelgrpc"
-	otelCodes "go.opentelemetry.io/otel/codes"
-	"golang.org/x/text/language"
-	"google.golang.org/genproto/googleapis/rpc/errdetails"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/reflection"
 	"google.golang.org/grpc/status"
-	"hwlocale"
 	"hwutil"
 	"net"
-	"reflect"
-	"sort"
-	"strconv"
-	"strings"
 	"telemetry"
 )
 
@@ -56,16 +46,13 @@ type organizationIDKey struct{}
 func StartNewGRPCServer(ctx context.Context, addr string, registerServerHook func(*daprd.Server)) {
 	log := zlog.Ctx(ctx)
 
-	chain := DefaultInterceptorChain()
-	statsHandler := grpc.StatsHandler(otelgrpc.NewServerHandler())
-
 	listener, err := net.Listen("tcp", addr)
 	if err != nil {
 		log.Fatal().Str("addr", addr).Err(err).Send()
 	}
 
 	// dapr/grpc service
-	service := daprd.NewServiceWithListener(listener, chain, statsHandler).(*daprd.Server)
+	service := daprd.NewServiceWithListener(listener, DefaultServerOptions()...).(*daprd.Server)
 	server := service.GrpcServer()
 
 	if err := service.AddHealthCheckHandler("", func(ctx context.Context) error {
@@ -79,7 +66,7 @@ func StartNewGRPCServer(ctx context.Context, addr string, registerServerHook fun
 
 	if Mode == DevelopmentMode {
 		reflection.Register(server)
-		log.Warn().Msg("grpc reflection enabled")
+		log.Warn().Msg("gRPC reflection enabled")
 	}
 
 	interrupted, err := hwutil.RunUntilInterrupted(ctx, func() error {
@@ -105,30 +92,46 @@ func StartNewGRPCServer(ctx context.Context, addr string, registerServerHook fun
 	Shutdown(nil)
 }
 
-// DefaultInterceptors returns the slice of default Interceptors for the GRPC service
-// Also see DefaultInterceptorChain
+// DefaultUnaryInterceptors returns the slice of default interceptors for unary gRPC calls
 //
-//	chain := grpc.ChainUnaryInterceptor(common.DefaultInterceptors()...)
-func DefaultInterceptors() []grpc.UnaryServerInterceptor {
+//	chain := grpc.ChainUnaryInterceptor(common.DefaultUnaryInterceptors()...)
+func DefaultUnaryInterceptors(metrics *prometheusGrpcProvider.ServerMetrics) []grpc.UnaryServerInterceptor {
+	return []grpc.UnaryServerInterceptor{
+		metrics.UnaryServerInterceptor(),
+		hwgrpc.UnaryLoggingInterceptor,
+		hwgrpc.UnaryErrorQualityControlInterceptor,
+		hwgrpc.UnaryLocaleInterceptor,
+		authUnaryInterceptor,
+		hwgrpc.UnaryValidateInterceptor,
+		hwgrpc.UnarySpanInterceptor,
+	}
+}
 
+// DefaultStreamInterceptors returns the slice of default interceptors for stream gRPC calls
+//
+//	chain := grpc.ChainStreamInterceptor(common.DefaultStreamInterceptors()...)
+func DefaultStreamInterceptors(metrics *prometheusGrpcProvider.ServerMetrics) []grpc.StreamServerInterceptor {
+	return []grpc.StreamServerInterceptor{
+		metrics.StreamServerInterceptor(),
+		hwgrpc.StreamLoggingInterceptor,
+		hwgrpc.StreamErrorQualityControlInterceptor,
+		hwgrpc.StreamLocaleInterceptor,
+		// authUnaryInterceptor,
+		hwgrpc.StreamValidateInterceptor,
+		hwgrpc.StreamSpanInterceptor,
+	}
+}
+
+func DefaultServerOptions() []grpc.ServerOption {
 	// register new metrics collector with prometheus
 	metrics := prometheusGrpcProvider.NewServerMetrics()
 	telemetry.PrometheusRegistry().MustRegister(metrics)
 
-	interceptors := []grpc.UnaryServerInterceptor{
-		metrics.UnaryServerInterceptor(),
-		loggingUnaryInterceptor,
-		errorQualityControlInterceptor,
-		localeInterceptor,
-		authUnaryInterceptor,
-		validateUnaryInterceptor,
-		handlerSpanInterceptor,
-	}
-	return interceptors
-}
+	unaryInterceptorChain := grpc.ChainUnaryInterceptor(DefaultUnaryInterceptors(metrics)...)
+	streamInterceptorChain := grpc.ChainStreamInterceptor(DefaultStreamInterceptors(metrics)...)
+	statsHandler := grpc.StatsHandler(otelgrpc.NewServerHandler())
 
-func DefaultInterceptorChain() grpc.ServerOption {
-	return grpc.ChainUnaryInterceptor(DefaultInterceptors()...)
+	return []grpc.ServerOption{unaryInterceptorChain, streamInterceptorChain, statsHandler}
 }
 
 func authUnaryInterceptor(ctx context.Context, req interface{}, info *grpc.UnaryServerInfo, next grpc.UnaryHandler) (interface{}, error) {
@@ -330,378 +333,12 @@ func GetOrganizationID(ctx context.Context) (uuid.UUID, error) {
 	}
 }
 
-// localeInterceptor parses the Accept-Language header.
-// Also see hwlocale.WithLocales, hwlocale.GetLocalesTags and hwlocale.GetLocalesStrings
-func localeInterceptor(ctx context.Context, req interface{}, info *grpc.UnaryServerInfo, next grpc.UnaryHandler) (interface{}, error) {
-	log := zlog.Ctx(ctx)
-
-	metaData := metadata.ExtractIncoming(ctx)
-
-	langHeader := metaData.Get("accept-language")
-	tags, ok := parseLocales(langHeader)
-	if !ok {
-		log.Warn().Str("langHeader", langHeader).Msg("invalid accept-language header received")
+// OrganizationIDFromMD retrieves the user defined organizationID
+// from the metadata of the request
+func OrganizationIDFromMD(ctx context.Context) (string, error) {
+	val := metadata.ExtractIncoming(ctx).Get("X-Organization")
+	if val == "" {
+		return "", status.Errorf(codes.Unauthenticated, "organization header missing")
 	}
-
-	return next(hwlocale.WithLocales(ctx, tags), req)
-}
-
-// parseLocales parses an HTTP Accept-Language Header Value into a descending-sorted language.Tag slice wrt. priority.
-// The bool indicates success.
-// See: https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Accept-Language
-// and: https://datatracker.ietf.org/doc/html/rfc2616#section-14.4
-func parseLocales(langHeader string) ([]language.Tag, bool) {
-	type langT struct {
-		tag language.Tag
-		q   float32
-	}
-	langs := make([]langT, 0)
-
-	parts := strings.Split(langHeader, ",")
-	for _, part := range parts {
-		qWeightParts := strings.Split(part, ";q=")
-		var q float32 = 1.0 // default q-weight
-		lang := strings.TrimSpace(qWeightParts[0])
-
-		if len(qWeightParts) != 1 {
-			if parsed, err := strconv.ParseFloat(strings.TrimSpace(qWeightParts[1]), 32); err != nil {
-				// invalid header, stop parsing
-				return []language.Tag{}, false
-			} else {
-				q = float32(parsed)
-			}
-		}
-		if tag, err := language.Parse(lang); err != nil {
-			continue // maybe we just don't know the language
-		} else {
-			langs = append(langs, langT{tag: tag, q: q})
-		}
-	}
-
-	// no priority queue needed, as we deal with small lengths
-	sort.SliceStable(langs, func(i, j int) bool {
-		return langs[i].q > langs[j].q
-	})
-	return hwutil.Map(langs, func(lang langT) language.Tag {
-		return lang.tag
-	}), true
-}
-
-// handlerSpanInterceptor opens and closes a span around the next in the chain
-// it should only be used as the last element in the interceptor chain before the handler
-func handlerSpanInterceptor(ctx context.Context, req interface{}, info *grpc.UnaryServerInfo, next grpc.UnaryHandler) (interface{}, error) {
-	ctx, span, _ := telemetry.StartSpan(ctx, "handler")
-	res, err := next(ctx, req)
-	if err != nil {
-		span.SetStatus(otelCodes.Error, err.Error())
-	}
-	span.End()
-	return res, err
-}
-
-// errorQualityControlInterceptor logs violations to https://cloud.google.com/apis/design/errors#error_payloads
-func errorQualityControlInterceptor(ctx context.Context, req interface{}, info *grpc.UnaryServerInfo, next grpc.UnaryHandler) (interface{}, error) {
-	res, err := next(ctx, req)
-
-	// no error, no error quality to control
-	if err == nil {
-		return res, err
-	}
-
-	log := zlog.Ctx(ctx)
-
-	statusError, ok := status.FromError(err)
-	if !ok {
-		log.Warn().
-			Err(err).
-			Str("type", reflect.TypeOf(err).String()).
-			Msg("non-status error was returned")
-		return res, NewStatusError(ctx, codes.Internal, err.Error(), locale.GenericError(ctx))
-	}
-
-	hasLocalizedMessage := false
-	hasBadRequest := false
-	hasPreconditionFailure := false
-	hasErrorInfo := false
-	hasResourceInfo := false
-	hasQuotaFailure := false
-
-	for _, detail := range statusError.Details() {
-		if _, ok := detail.(*errdetails.LocalizedMessage); ok {
-			hasLocalizedMessage = true
-		}
-		if _, ok := detail.(*errdetails.BadRequest); ok {
-			hasBadRequest = true
-		}
-		if _, ok := detail.(*errdetails.PreconditionFailure); ok {
-			hasPreconditionFailure = true
-		}
-		if _, ok := detail.(*errdetails.ErrorInfo); ok {
-			hasErrorInfo = true
-		}
-		if _, ok := detail.(*errdetails.ResourceInfo); ok {
-			hasResourceInfo = true
-		}
-		if _, ok := detail.(*errdetails.QuotaFailure); ok {
-			hasQuotaFailure = true
-		}
-	}
-
-	if !hasLocalizedMessage {
-		log.Warn().Err(err).Msg("status error does not have a LocalizedMessage")
-
-		var err2 error
-		statusError, err2 = statusError.WithDetails(LocalizedMessage(ctx, locale.GenericError(ctx)))
-		if statusError != nil {
-			err = statusError.Err()
-		} else {
-			log.Error().
-				Err(err2).
-				Msg("there was an error while creating the generic fallback statusError")
-			return res, err // respond with original error
-		}
-	}
-
-	switch statusError.Code() {
-	case codes.InvalidArgument:
-		fallthrough
-	case codes.OutOfRange:
-		if !hasBadRequest {
-			log.Warn().
-				Str("code", statusError.Code().String()).
-				Msg("status errors with this code should have a BadRequest detail, but none found")
-		}
-	case codes.FailedPrecondition:
-		if !hasPreconditionFailure {
-			log.Warn().
-				Str("code", statusError.Code().String()).
-				Msg("status errors with this code should have a PreconditionFailure detail, but none found")
-		}
-	case codes.Unauthenticated:
-		fallthrough
-	case codes.PermissionDenied:
-		fallthrough
-	case codes.Aborted:
-		if !hasErrorInfo {
-			log.Warn().
-				Str("code", statusError.Code().String()).
-				Msg("status errors with this code should have a ErrorInfo detail, but none found")
-		}
-	case codes.NotFound:
-		fallthrough
-	case codes.AlreadyExists:
-		if !hasResourceInfo {
-			log.Warn().
-				Str("code", statusError.Code().String()).
-				Msg("status errors with this code should have a ResourceInfo detail, but none found")
-		}
-	case codes.ResourceExhausted:
-		if !hasQuotaFailure {
-			log.Warn().
-				Str("code", statusError.Code().String()).
-				Msg("status errors with this code should have a QuotaFailure detail, but none found")
-		}
-	}
-
-	return res, err
-}
-
-func validateUnaryInterceptor(ctx context.Context, req interface{}, info *grpc.UnaryServerInfo, next grpc.UnaryHandler) (res interface{}, err error) {
-	ctx, span, _ := telemetry.StartSpan(ctx, "validate_interceptor")
-	defer func() {
-		if err != nil {
-			span.SetStatus(otelCodes.Error, err.Error())
-		}
-		span.End()
-	}()
-
-	log := zlog.Ctx(ctx)
-
-	err = hwutil.Validate(req)
-
-	// no error, go to next in chain
-	if err == nil {
-		log.Info().Msg("input validation ok")
-		return next(ctx, req)
-	}
-
-	// validate either returns an InvalidValidationError (in case there was an issue with the setup)
-	// or ValidationErrors (in case the input has validation issues)
-
-	var internalErr *validator.InvalidValidationError
-
-	if errors.As(err, &internalErr) {
-		return nil, NewStatusError(ctx, codes.Internal, err.Error(), locale.GenericError(ctx))
-	}
-
-	var valErrs validator.ValidationErrors
-
-	if errors.As(err, &valErrs) {
-		br := &errdetails.BadRequest{FieldViolations: make([]*errdetails.BadRequest_FieldViolation, 0)}
-
-		for _, valErr := range valErrs {
-			br.FieldViolations = append(br.FieldViolations, &errdetails.BadRequest_FieldViolation{
-				Field:       valErr.Field(), // TODO: use json key
-				Description: validateFieldErrDescription(ctx, valErr),
-			})
-		}
-
-		return nil, NewStatusError(
-			ctx,
-			codes.InvalidArgument,
-			err.Error(),
-			locale.GenericInvalidArgsError(ctx, len(valErrs)),
-			br,
-		)
-	}
-
-	log.Error().
-		Err(err).
-		Str("type", reflect.TypeOf(err).String()).
-		Msg("validate returned unexpected error")
-
-	return nil, NewStatusError(ctx, codes.Internal, err.Error(), locale.GenericError(ctx))
-}
-
-func validateFieldErrDescription(ctx context.Context, fieldErr validator.FieldError) string {
-	log := zlog.Ctx(ctx)
-	var l hwlocale.Locale
-	switch fieldErr.Tag() {
-	case "required":
-		l = locale.RequiredError(ctx)
-	// TODO: add more tags
-	default:
-		log.Warn().Str("tag", fieldErr.Tag()).Msg("tag unhandled, falling back to InvalidFieldError description")
-		l = locale.InvalidFieldError(ctx)
-	}
-
-	return hwlocale.Localize(ctx, l)
-}
-
-func loggingUnaryInterceptor(ctx context.Context, req interface{}, info *grpc.UnaryServerInfo, next grpc.UnaryHandler) (interface{}, error) {
-	ctx, span, log := telemetry.StartSpan(ctx, "logging_interceptor")
-	defer span.End()
-
-	metaData := metadata.ExtractIncoming(ctx)
-
-	// Add request information
-	builder := log.
-		With().
-		Str("handler", info.FullMethod)
-
-	omitBody := false
-
-	// additional information for pub/sub events
-	// TODO: make sure this is picked up in otel propagation
-	if req, ok := req.(*runtime.TopicEventRequest); ok {
-		if traceparent, ok := req.GetExtensions().Fields["traceparent"]; ok {
-			builder = builder.Str("traceparent", traceparent.GetStringValue())
-		}
-		builder = builder.Str("eventID", req.Id)
-
-		// at this point in the chain we have no control about what data may be logged for events,
-		// so we can't log anything for privacy and/or legal reasons, the event handler can log though
-		omitBody = true
-	}
-
-	// this is the logger that should be used for this request
-	log = builder.Logger()
-
-	// context, with new logger attached
-	ctx = log.WithContext(ctx)
-
-	log.Trace().Interface("metaData", redactMetadata(metaData)).Send()
-
-	logBody := req
-
-	// custom map instead, if implemented
-	if loggable, ok := req.(telemetry.Loggable); ok {
-		logBody = loggable.LoggableFields()
-	}
-	if !omitBody {
-		log.Debug().Interface("body", logBody).Type("bodyType", req).Send()
-	}
-
-	// Call next in chain
-	res, err := next(ctx, req)
-
-	// We are back! Now evaluate response
-
-	if err != nil {
-		statusError := status.Convert(err)
-		code := statusError.Code().String()
-		message := statusError.Message()
-		details := statusError.Details()
-
-		log.
-			WithLevel(resolveLogLevelForError(err)).
-			Err(err).
-			Str("code", code).
-			Interface("details", details).
-			Msg(message)
-	} else {
-		log.
-			Info().
-			Str("code", codes.OK.String()).
-			Msg("Ok")
-	}
-
-	// pass results back up the interceptor chain
-	return res, err
-}
-
-// resolveLogLevelForError is primary used to raise the log level on specific errors
-func resolveLogLevelForError(err error) zerolog.Level {
-	statusError := status.Convert(err)
-	if statusError.Code() == codes.Internal {
-		return zerolog.ErrorLevel
-	}
-
-	return zerolog.WarnLevel
-}
-
-func redactMetadata(m metadata.MD) metadata.MD {
-	if arr := m["authorization"]; arr != nil {
-		for i := range arr {
-			arr[i] = telemetry.OmitAll(arr[i])
-		}
-	}
-	return m
-}
-
-// NewStatusError builds a new Status Error
-// - TODO: add Error Reason
-// - msg is developer-facing and must be in english
-// - for a user-facing message use locale
-// - additional details can be added
-// For more information on details see: https://cloud.google.com/apis/design/errors#error_details
-// and https://cloud.google.com/apis/design/errors#error_payloads
-func NewStatusError(ctx context.Context, code codes.Code, msg string, locale hwlocale.Locale, details ...proto.Message) error {
-	log := zlog.Ctx(ctx)
-	statusNoDetails := status.New(code, msg)
-
-	if locale.Bundle == nil || locale.Config == nil {
-		log.Warn().Msg("creating Status Error without valid locale")
-	} else {
-		details = append(details, LocalizedMessage(ctx, locale))
-	}
-
-	s, err := statusNoDetails.WithDetails(details...)
-	if err != nil {
-		log.Warn().
-			Err(err).
-			Msg("could not add details to status error")
-		return statusNoDetails.Err()
-	}
-	return s.Err()
-}
-
-// LocalizedMessage returns a LocalizedMessage Error Detail as per https://cloud.google.com/apis/design/errors#error_details
-// Also see NewStatusError, which constructs a LocalizedMessage for you
-func LocalizedMessage(ctx context.Context, locale hwlocale.Locale) *errdetails.LocalizedMessage {
-	str, tag := hwlocale.LocalizeWithTag(ctx, locale)
-	return &errdetails.LocalizedMessage{
-		Locale:  tag.String(),
-		Message: str,
-	}
+	return val, nil
 }
