+++ libs/common/hwgrpc/export_test.go
@@ -1,4 +1,4 @@
-package common
+package hwgrpc
 
 // apparently this is needed
 var ParseLocales = parseLocales
