+++ libs/common/hwgrpc/hwgrpc_test.go
@@ -1,7 +1,7 @@
-package common_test
+package hwgrpc_test
 
 import (
-	"common"
+	"common/hwgrpc"
 	"golang.org/x/text/language"
 	"testing"
 )
@@ -21,7 +21,7 @@ func arrayEq(t *testing.T, expected []language.Tag, parsed []language.Tag) {
 func TestParseLocales(t *testing.T) {
 	const mdnExample = "fr-CH, fr;q=0.9, en;q=0.8, de;q=0.7, *;q=0.5"
 	expected := []language.Tag{language.MustParse("fr-CH"), language.French, language.English, language.German}
-	parsed, ok := common.ParseLocales(mdnExample)
+	parsed, ok := hwgrpc.ParseLocales(mdnExample)
 	if !ok {
 		t.Error("failed to parse")
 		return
@@ -32,7 +32,7 @@ func TestParseLocales(t *testing.T) {
 func TestParseLocalesReordered(t *testing.T) {
 	const mdnExample = "fr-CH, de;q=0.7, en;q=0.8, fr;q=0.9, *;q=0.5"
 	expected := []language.Tag{language.MustParse("fr-CH"), language.French, language.English, language.German}
-	parsed, ok := common.ParseLocales(mdnExample)
+	parsed, ok := hwgrpc.ParseLocales(mdnExample)
 	if !ok {
 		t.Error("failed to parse")
 		return
@@ -43,7 +43,7 @@ func TestParseLocalesReordered(t *testing.T) {
 func TestParseLocalesSimple(t *testing.T) {
 	const mdnExample = "de"
 	expected := []language.Tag{language.German}
-	parsed, ok := common.ParseLocales(mdnExample)
+	parsed, ok := hwgrpc.ParseLocales(mdnExample)
 	if !ok {
 		t.Error("failed to parse")
 		return
