+++ libs/hwes/event_test.go
@@ -23,3 +23,19 @@ func TestEventWithUserID(t *testing.T) {
 		t.Error("event does not have the correct CommitterUserID")
 	}
 }
+
+func TestEventWithOrganizationID(t *testing.T) {
+	ctx := context.Background()
+	u := uuid.New()
+	e := hwes.Event{}
+
+	ctx = common.ContextWithOrganizationID(ctx, u)
+
+	if err := e.SetOrganizationFromCtx(ctx); err != nil {
+		t.Error(err)
+	}
+
+	if *e.OrganizationID != u {
+		t.Error("event does not have the correct OrganizationID")
+	}
+}
