+++ services/property-svc/internal/property-view/api/grpc_test.go
@@ -34,7 +34,7 @@ func server() (context.Context, pb.PropertyViewsServiceClient, *hwes_test.Aggreg
 	ctx := common.Setup("property-svc", "test", common.WithFakeAuthOnly())
 
 	// Start Server
-	grpcServer := grpc.NewServer(common.DefaultInterceptorChain())
+	grpcServer := grpc.NewServer(common.DefaultServerOptions()...)
 	pb.RegisterPropertyViewsServiceServer(grpcServer, grpcService)
 	conn, closer := common_test.StartGRPCServer(ctx, grpcServer)
